# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/availableHours/": {
            "get": {
                "description": "List all the hours remaining for creaiting an entrie on the schedule",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "titulacion de las horas a obtener",
                        "name": "titulacion",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "curso de las horas a obtener",
                        "name": "curso",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "grupo de las horas a obtener",
                        "name": "grupo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.SchedulerDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorHttp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorHttp"
                        }
                    }
                }
            }
        },
        "/listDegrees/": {
            "get": {
                "description": "List all degrees' descriptions avaiable, it do not require any parameter",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ListDegreesDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorHttp"
                        }
                    }
                }
            }
        },
        "/ping/": {
            "get": {
                "description": "Response \"pong\" if the server is currrently available",
                "produces": [
                    "text/plain"
                ],
                "responses": {
                    "200": {
                        "description": "Returns \"pong\" "
                    }
                }
            }
        },
        "/updateScheduler/": {
            "post": {
                "description": "The request will erase the current scheduler an create one new with\nthe requested entries for the scheduler. The entry will be definied by the initial hour\nand the ending hour, adintional info must be indicated depending of the kind of hours\nthe kinds of subject hours are:\n- Theorical = 1\n- Practices = 2\n- Exercises = 3",
                "produces": [
                    "text/plain"
                ],
                "parameters": [
                    {
                        "description": "Entry to create",
                        "name": "entry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.EntryDTO"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receive the date of the latests entry modification with format dd/mm/aaaa"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.AvailableHours": {
            "type": "object",
            "properties": {
                "max": {
                    "type": "integer"
                },
                "remaining": {
                    "type": "integer"
                },
                "subject": {
                    "$ref": "#/definitions/domain.Subject"
                }
            }
        },
        "domain.DegreeDescription": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "years": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.YearDescription"
                    }
                }
            }
        },
        "domain.Entry": {
            "type": "object",
            "properties": {
                "end": {
                    "$ref": "#/definitions/domain.Hour"
                },
                "group": {
                    "type": "string"
                },
                "init": {
                    "$ref": "#/definitions/domain.Hour"
                },
                "room": {
                    "$ref": "#/definitions/domain.Room"
                },
                "subject": {
                    "$ref": "#/definitions/domain.Subject"
                },
                "week": {
                    "type": "string"
                },
                "weekday": {
                    "type": "integer"
                }
            }
        },
        "domain.Hour": {
            "type": "object"
        },
        "domain.Room": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.Subject": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.YearDescription": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "integer"
                }
            }
        },
        "handlers.EntryDTO": {
            "type": "object",
            "properties": {
                "endHour": {
                    "type": "integer"
                },
                "endMin": {
                    "type": "integer"
                },
                "grupo": {
                    "type": "string"
                },
                "initHour": {
                    "type": "integer"
                },
                "initMin": {
                    "type": "integer"
                },
                "kind": {
                    "type": "integer"
                },
                "room": {
                    "type": "string"
                },
                "semana": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "weekday": {
                    "type": "integer"
                }
            }
        },
        "handlers.ErrorHttp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.ListDegreesDTO": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.DegreeDescription"
                    }
                }
            }
        },
        "handlers.SchedulerDTO": {
            "type": "object",
            "properties": {
                "availableHours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AvailableHours"
                    }
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Entry"
                    }
                }
            }
        }
    }
}